SaaS Subscription Churn Analysis
Organisations enjoy the retention of large customer bases as that drives revenue and profit. Churn analysis is important as it allows organisations to identify trends that influence churn. Here we look at a dataset of a SaaS and the various factors that affect Churn

Dataset: https://www.kaggle.com/datasets/safrin03/predictive-analytics-for-customer-ch urn-dataset